Fixed differences in data types in pojo classes and entities

+++ src/main/java/ua/epam/javard/weatherApp/entity/WeatherDetailsEntity.java
@@ -38,7 +38,7 @@ public class WeatherDetailsEntity {
     private String windDirFull;
 
     @Column(name = "wind_dir_degrees", nullable = false)
-    private Double windDirDegrees;
+    private Long windDirDegrees;
 
     @Column(name = "wind_speed", nullable = false)
     private Double windSpeed;
@@ -50,6 +50,6 @@ public class WeatherDetailsEntity {
     private Double pressure;
 
     @Column(name = "humidity", nullable = false)
-    private Double humidity;
+    private Long humidity;
 
 }
